+++ Labo1/add.h
@@ -18,7 +18,7 @@ Compilateur     : Mingw-w64 g++ 8.1.0
 
 class Add : public Operation{
 public:
-    unsigned apply(unsigned a, unsigned b);
+    long long apply(long long a, long long b);
 };
