+++ Labo1/multiply.h
@@ -16,7 +16,7 @@ Compilateur     : Mingw-w64 g++ 8.1.0
 
 class Multiply : public Operation{
 public:
-    unsigned apply(unsigned a, unsigned b);
+    long long apply(long long a, long long b);
 };
